+++ module/CLI/test/Command/Visit/LocateVisitsCommandTest.php
@@ -47,7 +47,7 @@ class LocateVisitsCommandTest extends TestCase
 
         $locker = $this->createMock(Lock\LockFactory::class);
         $this->lock = $this->createMock(Lock\SharedLockInterface::class);
-        $locker->method('createLock')->with($this->isType('string'), 600.0, false)->willReturn($this->lock);
+        $locker->method('createLock')->with($this->isString(), 600.0, false)->willReturn($this->lock);
 
         $command = new LocateVisitsCommand($this->visitService, $this->visitToLocation, $locker);